code/requirements.txt
pyopengl
numpy
glfw
imgui